refactor(web): ♻️ misc. tidy
