+++ web/components/header/mode-toggle.tsx
@@ -1,9 +1,9 @@
+// https://ui.shadcn.com/docs/dark-mode/next
 "use client";
 
 import * as React from "react";
 import { MoonIcon, SunIcon } from "@radix-ui/react-icons";
 import { useTheme } from "next-themes";
-
 import { Button } from "@/components/ui/button";
 import {
   DropdownMenu,
