+++ web/components/theme-provider.tsx
@@ -1,3 +1,4 @@
+// https://ui.shadcn.com/docs/dark-mode/next
 "use client";
 
 import * as React from "react";
