+++ web/components/ui/data-table.tsx
@@ -1,3 +1,4 @@
+// https://ui.shadcn.com/docs/components/data-table
 "use client";
 
 import {
@@ -9,7 +10,6 @@ import {
   getSortedRowModel,
   useReactTable,
 } from "@tanstack/react-table";
-
 import {
   Table,
   TableBody,
